campus-safety/backend/requirements.txt
flask>=3.0.3
flask-socketio>=5.3.6
flask-cors>=4.0.1
opencv-python>=4.10.0
ultralytics>=8.3.0
dashscope>=1.20.11
numpy>=1.26.4
Pillow>=10.3.0
gevent>=24.2.1
gevent-websocket>=0.10.1
python-dotenv>=1.0.1
requests>=2.32.3
